requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.3
pandas==2.2.1
sqlalchemy==2.0.27
scikit-learn==1.4.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9